--- try-3 best/test.py
@@ -1,5 +1,0 @@
-import requests
-
-resp = requests.post("http://127.0.0.1:5000/predict", files={'file': open('bottle2.jpg', 'rb')})
-
-print(resp.json())
