+++ try-3 photo based/requirement.txt
@@ -1,0 +1,8 @@
+Flask
+gunicorn
+tensorflow
+numpy
+pillow
+Keras-Applications
+Keras-Preprocessing
+opencv-python
