+++ try-4 url based/requirement.txt
@@ -1,0 +1,8 @@
+Flask
+gunicorn
+tensorflow
+numpy
+pillow
+Keras-Applications
+keras-preprocessing
+opencv-python
